requirements.txt
Flask
google-cloud-storage
gunicorn
    